requirements.txt
streamlit==1.29.0
moviepy==1.0.3
openai-whisper
librosa==0.10.1
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.3
nltk==3.8.1
yt-dlp
plotly==5.17.0
wordcloud==1.9.3
Pillow==9.5.0
torchaudio
torch
